Add embed link test to main_page_test.py
